+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 - the DTO code-generator now has a parameter that allows generating C# records: ```--model-type ImmutableClass|Record``` or ```<SqModelGenType>ImmutableClass|Record</SqModelGenType>```;
 - "CheckExistenceBy" in the Insert data builder what adds WHERE EXISTS(...) to a Insert source to avoid duplicates inserting;
 - ExistsIn ...
+- QueryPage extension for OffsetFetch queries.
 ### Bugfix
 - When some column(s) in values constructor contains only nulls, sqexpress now adds an explicit type cast for the first cell e.g. ```CAST(NULL as int)```
 - **MergeDataInto** now allows only keys mapping unless **WhenMatchThenUpdate** is defined (without additional updates)
+- Derived tables now can have 0 declared columns
 
 # 0.3.0.0
 ### New Features